Updated quest code to work in testing level
Only working male level, not in female level. Main gui components are
affecting the quest gui in some way. We will need to adjust the main gui
so that it works properly with everything else as it appears to be done
in a different way than the other gui components.