Completed controller for player and enemies (12 states, each with 8 transitions)
Controller has 12 states, and each of these states has 8 transitions to
move to various other states. 3 variables (1 int, 2 bool) are used to
keep track of which state the controller is in. With 96 total
transitions, we will need to ensure that these variables are properly
updated throughout scripts for player and enemies.